Convert typescript utils into javascript
